+++ apps/portal/locale/he.json
@@ -132,9 +132,9 @@
       },
       "scoreboard": {
         "rank": "דירוג",
-        "best-score": "ניקוד הטוב ביותר",
+        "best-score": "ניקוד מירבי",
         "match": "מקצה",
-        "match-scores": "ניקוד מקצים",
+        "match-scores": "תוצאות מקצים",
         "no-data": "אין נתוני לוח תוצאות זמינים"
       },
       "field-schedule": {
